Related work items: #12947, #12948, #12950, #12951, #13098, #13099, #13100, #13101, #13104, #13106, #13107, #13108, #13109, #14889, #14950

+++ dotnet/corecomponents/net.atos.daf.ct2.reports/entity/ReportDetails.cs
@@ -20,7 +20,7 @@ namespace net.atos.daf.ct2.reports.entity
         public char Type { get; set; }
         public char? ChartType { get; set; }
         public string ThresholdType { get; set; }
-        public long ThresholdValue { get; set; }
+        public double ThresholdValue { get; set; }
     }
     public class ReportListedParamaters
     {
